app/CC/CC-server/pyproject.toml
[project]
name = "cc-server"
version = "0.1.0"
description = "CC-server — Call Center orchestrator, WebRTC signaling relay, A2A flow coordinator"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "a2a-sdk[http-server]>=0.2.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "websockets>=13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
